Chore: Update yargs-parser due to a security update
The security vulnerability is described here: https://github.com/microsoft/axe-sarif-converter/network/alert/yarn.lock/yargs-parser/open

The change was to call

- yarn remove semantic-release
- yarn add --dev semantic-release

In addition, a resolution was added to package.json to force one of our dependencies to use a later version of yargs. We specify yargs instead of yargs-parser so that the two packages have a better chance at compatibility.