Update to the latest version of a package with a possible security issue
This came in as a cg warning. I don't think we use the package in question in a way that would make us vulnerable, but updating the package to the latest version seemed like the wise choice of action.

Warning: The package npm-user-validate prior to version 1.0.1 is vulnerable to REDoS. The regex that validates a user's email took exponentially longer to process
input strings that begin with the '@' character.