chore: try out Renovate instead of Dependabot for dependency updates (#685)
#### Details

This PR is an initial attempt to trial using [Renovate](https://docs.renovatebot.com/) instead of Dependabot for automated dependency updates.

This PR uses renovate/github-action to implement a self-hosted Renovate configuration. At a high level, this works by using a GitHub workflow that runs on a schedule every 15 minutes to run a "renovate" command, which queries for any new dependency updates and creates/updates/merges any required dependency update PRs.

For the purposes of the trial experiment, this workflow is currently set up just in this local repo. If we like how this works in `axe-sarif-converter`, I think the better long-term solution would be to create a separate `microsoft/accessibility-insights-renovate-config` repo to contain a single common worker workflow + config preset, and to have one worker covering updates for all of our repos rather than repeating this workflow/config/etc in each of our repos.

This initial PR sets `"dryRun": true` in `renovate-self-host-config.json5`; once this is merged and I've checked over the dry run log file, a separate PR will remove that config line to enable the workflow to actually update PRs.

##### Motivation

The primary motivations for trying this are:

* Renovate's `stabilityDays` config option would act as partial mitigation for issues like the ua-parser-js compromise we saw a few weeks ago (it doesn't solve issues where a package's maintainer is malicious, but it mitigates issues where a non-malicious maintainer's publishing credentials are temporarily compromised)
* Renovate supports more package managers than Dependabot does. Notably, it supports both pnpm and yarn v3; these are appealing because they both have a better version of the yarn v1 `resolutions` we currently use for certain security updates in transitive dependencies, which currently forces us to accept technical debt every time we have to use them.
* Renovate supports grouping related package updates together. Its default `config:base` preset includes grouping for many common monorepos (eg, gatsby, jest, babel), and it also supports user-defined groupings.
* Renovate supports granular automerge settings (eg, "automerge any @types/* packages with passing tests, but don't automerge any other packages")

##### Context

There are two major alternatives to the self-hosted GitHub Action configuration:

* The Renovate GitHub App: I dismissed this because it requires configuration at the level of the `microsoft` GitHub organization, which we didn't want to treat as a blocker for trial experiment purposes
* Self-hosting via an Azure VM running a cron job: I dismissed this because it seemed like it would require both slightly more cost and slightly more dev overhead (in terms of managing OS updates/etc) than the GitHub Actions version. That said, if performance turns out to be an issue, this would be a reasonable alternative.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [x] PR title respects [Conventional Commits](https://www.conventionalcommits.org) (starts with `fix:`, `feat:`, etc, and is suitable for user-facing release notes)
- [x] PR contains no breaking changes, **OR** description of both PR **and final merge commit** starts with `BREAKING CHANGE:`
- [n/a] (if applicable) Addresses issue: #0000
- [n/a] Added relevant unit tests for your changes
- [x] Ran `yarn precheckin`
- [n/a] Verified code coverage for the changes made