Adding required SECURITY.md file to the repo
After checking with the Github in Microsoft (GiM) team, we are being asked to add this file indicating Microsoft's security preferences to all open-source repos. This file is used to help the community understand the security policy and how to safely report security issues. GitHub uses the presence of this file to light-up security reminders and a link to the file. More information for internal reference can be found here: https://aka.ms/opensource