feat(axe-core-4.6): Update axe-core to 4.6.3 (#918)
#### Details
This PR updates the repo to use axe-core 4.6.3
<!-- Usually a sentence or two describing what the PR changes -->

##### Motivation
We are updating Accessibility Insights products to use axe-core 4.6.3
<!-- This can be as simple as "addresses issue #123" -->

##### Context
The first few commits represent the update to axe-core 4.6.3 and
updating the main tests for the repo. The last part of the PR (handled
by @shanisebarona) will update the wcag-link-data.test to account for
the fact that axe-core 4.6.3 references one success criterion from WCAG
2.2.
<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->
Yes. We considered updating the converter files and related wcag-link
components to account for two WCAG versions, but this proved more
complicated than what seemed reasonable based on the amount of support
it would provide to users. Instead, we've opted to not reference the 1
WCAG 2.2 criterion.
#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [X] PR title respects [Conventional
Commits](https://www.conventionalcommits.org) (starts with `fix:`,
`feat:`, etc, and is suitable for user-facing release notes)
- [X] PR contains no breaking changes, **OR** description of both PR
**and final merge commit** starts with `BREAKING CHANGE:`
- [N/A ] (if applicable) Addresses issue: #0000
- [X] Added relevant unit tests for your changes
- [ ] Ran `yarn precheckin`
- [ ] Verified code coverage for the changes made

---------

Co-authored-by: SB <shanisebarona@gmail.com>