chore: migrate to yarn v3 (#935)
#### Details

This PR migrates the repo from Yarn v1 to Yarn v3 for dependency
management. The main migration followed the steps in [Yarn's migration
guide](https://yarnpkg.com/getting-started/migration) for a
non-zero-installs configuration, similar to the migrations in
accessibility-insights-service and accessibility-insights-action.

Besides the required/mechanical steps, this PR also includes:

* Updating the resolutions used by this repo to use versioned
resolutions, similar to
https://github.com/microsoft/accessibility-insights-action/pull/1596
* Modifying the `test-resources` generator setup to work as a run script
from the root directory instead of using a separate `package.json` in
`src/test-resources/generator`. This avoided what would have otherwise
required a more complicated setup using either multiple checked in
copies of yarn or a more complicated workspace setup, and slightly
simplifies the steps for using the generator. I updated the README
instructions accordingly.
* The new `!dist/test-resources` entry in the `files` field of
`package.json` ensures that there is no change to the package
as-distributed because of this update.

As part of these changes, I verified manually that `yarn
generate-test-resources` regenerates the current versions of the test
resource files successfully.

##### Motivation

* Consistency with other repos
* Allows use of versioned resolutions (see
https://github.com/microsoft/accessibility-insights-action/pull/1596 for
why we want them)

##### Context

Similar PRs in other repos:

* https://github.com/microsoft/accessibility-insights-service/pull/2210
* https://github.com/microsoft/accessibility-insights-action/pull/1559

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [x] PR title respects [Conventional
Commits](https://www.conventionalcommits.org) (starts with `fix:`,
`feat:`, etc, and is suitable for user-facing release notes)
- [x] PR contains no breaking changes, **OR** description of both PR
**and final merge commit** starts with `BREAKING CHANGE:`
- [n/a] (if applicable) Addresses issue: #0000
- [n/a] Added relevant unit tests for your changes
- [x] Ran `yarn precheckin`
- [n/a] Verified code coverage for the changes made