fix: Add dummy location data to conform to SARIF spec (#988)
#### Details

As called out in #961, our SARIF doesn't have a fully formed location.
Since we don't have meaningful data for this, the "least bad" solution
was to insert dummy data for the `startLine` property. This PR includes
2 code changes, plus a bunch of refreshed SARIF files generated via the
CLI tool. I noticed that the refreshed version of
`basic-axe-3.2.2.sarif` is pretty-printed differently than what we had
before. I'm guessing that the tool shifted at some point and that this
is the first time the file has been regenerated with the current tool.

##### Motivation

Address #961 for better tool compatibility

##### Context

<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [x] PR title respects [Conventional
Commits](https://www.conventionalcommits.org) (starts with `fix:`,
`feat:`, etc, and is suitable for user-facing release notes)
- [x] PR contains no breaking changes, **OR** description of both PR
**and final merge commit** starts with `BREAKING CHANGE:`
- [x] (if applicable) Addresses issue: #961
- [n/a] Added relevant unit tests for your changes
- [x] Ran `yarn precheckin`
- [x] Verified code coverage for the changes made