fix: Change "Usage" heading in README to all caps (#1180)
#### Details
This is updating the text in the README for the Usage section to now
show "USAGE".
<!-- Usually a sentence or two describing what the PR changes -->

##### Motivation
Triggering a build to confirm that the new GH token is being consumed
properly
<!-- This can be as simple as "addresses issue #123" -->

##### Context
Quarterly DRI rotation
<!-- Are there any parts that you've intentionally left out-of-scope for
a later PR to handle? -->

<!-- Were there any alternative approaches you considered? What
tradeoffs did you consider? -->

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a"
in the checkbox -->
- [ ] PR title respects [Conventional
Commits](https://www.conventionalcommits.org) (starts with `fix:`,
`feat:`, etc, and is suitable for user-facing release notes)
- [ ] PR contains no breaking changes, **OR** description of both PR
**and final merge commit** starts with `BREAKING CHANGE:`
- [ ] (if applicable) Addresses issue: #0000
- [ ] Added relevant unit tests for your changes
- [ ] Ran `yarn precheckin`
- [ ] Verified code coverage for the changes made